selfAndBelow: true now means that the same permission will be checked on the parent before applying subsequent ACLs on the child(ren).
selfAndBelow also now supports 2 additional forms:

If it's the name of a permission, the parent ACL for that permission will be checked first regardless of the actual permission in the request.  The child ACLs will then be checked using the requested permission.

If selfAndBelow is a function, it will be passed the user, permission, and environment and evaluated.  If the result is true, the subsequent child ACLs will be checked.  Otherwise, permission is not granted.